refactor: notification 관련 의존성 방향 수정
