+++ backend/src/main/java/com/happy/friendogly/chatsocket/service/ChatSocketCommandService.java
@@ -17,7 +17,7 @@ import com.happy.friendogly.club.repository.ClubRepository;
 import com.happy.friendogly.exception.FriendoglyException;
 import com.happy.friendogly.member.domain.Member;
 import com.happy.friendogly.member.repository.MemberRepository;
-import com.happy.friendogly.notification.service.ChatNotificationService;
+import com.happy.friendogly.chatroom.service.ChatNotificationService;
 import java.time.LocalDateTime;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
