+++ backend/src/main/java/com/happy/friendogly/deveicetoken/dto/response/UpdateDeviceTokenResponse.java
@@ -1,4 +1,4 @@
-package com.happy.friendogly.notification.dto.response;
+package com.happy.friendogly.deveicetoken.dto.response;
 
 public record UpdateDeviceTokenResponse(String deviceToken) {
