+++ backend/src/main/java/com/happy/friendogly/footprint/service/FootprintCommandService.java
@@ -15,7 +15,6 @@ import com.happy.friendogly.footprint.dto.response.UpdateWalkStatusManualRespons
 import com.happy.friendogly.footprint.repository.FootprintRepository;
 import com.happy.friendogly.member.domain.Member;
 import com.happy.friendogly.member.repository.MemberRepository;
-import com.happy.friendogly.notification.service.FootprintNotificationService;
 import com.happy.friendogly.pet.domain.Pet;
 import com.happy.friendogly.pet.repository.PetRepository;
 import java.time.LocalDateTime;
